feat(infra): set "RemovalPolicy.SNAPSHOT" for the RDS instance
Previously `RemovalPolicy.RETAIN` was set to DB and when trying to destroy the
stack, the RDS instances will retain as expected, but it prevents the related
resources (e.g. subnet, security group) to be deleted, which leads to the
failure of stack destroy. To make sure that the prod stack can be destroyed
successfully, I choose to substitute `RemovalPolicy.SNAPSHOT` for the previous
one.